fix up PR comments
