+++ .github/workflows/fetchy.yml
@@ -29,6 +29,8 @@ jobs:
         run: npm ci
       - name: Run Fetchy
         run: npm run fetchy
+      - name: Run Tests
+        run: npm test
       - uses: EndBug/add-and-commit@v7
         with:
           default_author: github_actions
