+++ fetchy/vaccineDataExtractor.ts
@@ -1,5 +1,5 @@
 import { VaccineData } from "../shared/types/VaccineDataTypes";
-import { getVaccinationDataPerDhb } from "./extractors/vaccinationDataPerDhbExtrator";
+import { getVaccinationDataPerDhb } from "./extractors/vaccinationDataPerDhbExtractor";
 import { extractDailyUpdatedNzTime } from "./extractors/dailyUpdatedTimeExtractor";
 
 export default function extractVaccineData(rawHtml: string): VaccineData {
