+++ netlify.toml
@@ -1,5 +1,5 @@
 [build]
-  command = "npm run test && npm run build"
+  command = "npm run build"
   publish = ".next"
 
 [functions]
